delete unnecessary things
