+++ Cargo.toml
@@ -107,16 +107,8 @@ name="fourier"
 path = "src/fourier.rs"
 
 [[bin]]
-name="peace2"
-path = "src/peace2.rs"
-
-[[bin]]
 name="nblur"
 path = "src/nblur.rs"
-
-[[bin]]
-name="vortexmona"
-path = "src/vortexmona.rs"
 
 [[bin]]
 name="pixelrain"
@@ -141,10 +133,6 @@ path = "src/pare.rs"
 [[bin]]
 name="neuralnet"
 path = "src/neuralnet.rs"
-
-[[bin]]
-name="peace3"
-path = "src/peace3.rs"
 
 [[bin]]
 name="bhole"
